% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_manoeli.R
\name{run_manoeli}
\alias{run_manoeli}
\title{Run manoeli}
\usage{
run_manoeli(get.r.help.from.alessandro = FALSE)
}
\arguments{
\item{get.r.help.from.alessandro}{Logical for getting help from Alessandro on
using R.}
}
\value{
A few published papers and, possibly, a PhD degree.
}
\description{
Analyse a lot of data, publish a few papers and, possibly, earn a PhD degree.
}
\details{
There is no guarantee that the papers will be accepted, or that you
will get a PhD degree.
}

